requirements.txt
openai==1.30.1
requests==2.31.0
beautifulsoup4==4.12.3
pillow==10.3.0
schedule==1.2.1
python-dotenv==1.0.1

# Additional dependencies for Pinterest SaaS
pandas>=2.0.0
selenium>=4.15.0
chromedriver-autoinstaller>=0.6.0
webdriver-manager>=4.0.0
matplotlib>=3.7.0
numpy>=1.24.0
httpx>=0.25.0
loguru>=0.7.0